Refactoring to be be exclusively a plugin
Previously this tried to mix being a plugin with being an extension
(overwriting loader.fetch, loader.normalize) which caused lots of problems and inconsistent behavior. Now it is just a plugin that exports a translate hook that will transpile in development and compile an entire TypeScript program in production. This also removes lots of 404 errors because we
are not trying to fetch anything as a .ts file other than what steal
does automatically (as well as any relative imports from other .ts
files).